style(first steps into new styles): new design
