+++ client/src/components/AreaPanel/AreaForm/AreaForm.module.scss
@@ -1,41 +1,41 @@
 @import 'styles/main.scss';
-.formsContainer{
-    min-height: fit-content;
-    padding: 15px;
-    @include card;
-    margin-top: 20px;
-    background-color: $blue;
-    width:95% ;
-    align-items: center;
-    justify-content: center;
-    display: flex;
-    flex-direction: column;
-  h3{
-      font-family: $titleFont;
-      margin: 2px;
-    } 
+
+.formsContainer {
+  @include card;
+  min-height: fit-content;
+  padding: 15px;
+  margin-top: 20px;
+  background-color: $blue;
+  width: 95%;
+  align-items: center;
+  justify-content: center;
+  display: flex;
+  flex-direction: column;
+  h3 {
+    font-family: $titleFont;
+    margin: 2px;
+  }
 }
-input{
-    min-height: 2rem;
-    max-height: 3rem;
-    min-width: 12rem;
-    max-width: 15rem;
+input {
+  min-height: 2rem;
+  max-height: 3rem;
+  min-width: 12rem;
+  max-width: 15rem;
 }
-.isactive{
-display: flex;
-margin-top: 10px;
-align-items: center;
-justify-content: space-between;
-font-size: larger;
+.isactive {
+  display: flex;
+  margin-top: 10px;
+  align-items: center;
+  justify-content: space-between;
+  font-size: larger;
 }
 
-.btn{
-    border-radius: $borderRadius;
-    border: none;
-    margin-bottom: 10px;
-    margin-top: 15px;
-    box-shadow: 4px 4px 10px #666;
-    cursor: pointer;
-    background-color: #dfc884;
-    
+.btn {
+  border-radius: $borderRadius;
+  border: none;
+  margin-bottom: 10px;
+  margin-top: 15px;
+  box-shadow: 4px 4px 10px #666;
+  cursor: pointer;
+  background-color: #dfc884;
 }
