+++ client/src/components/AreaPanel/AreaPanel.module.scss
@@ -1,28 +1,29 @@
 @import 'styles/main';
-.container{
-    max-height: 100vh;
-} 
-.title{
-    max-width: 90%;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-around;
-    margin: 10px;
-    button{
-        @include button-hover;
-        background-color: #fce08d;
-        border: none;
-        max-width: 2,5rem;
-        min-width: 2.5rem;
-        max-height: 2.5rem;
-        min-height: 2.5rem;
-        border-radius: $borderRadius;
-        font-weight: 1000%;
-        font-size:x-large;
-        color: rgb(0, 61, 0);
-        box-shadow:  6px 6px 7px rgb(168, 168, 168);
-        cursor: pointer;
-        outline:none;
-    }
+.container {
+  max-height: 100vh;
+  grid-area: areas;
+}
+.title {
+  max-width: 90%;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-around;
+  margin: 10px;
+  button {
+    @include button-hover;
+    background-color: #fce08d;
+    border: none;
+    max-width: 2, 5rem;
+    min-width: 2.5rem;
+    max-height: 2.5rem;
+    min-height: 2.5rem;
+    border-radius: $borderRadius;
+    font-weight: 1000%;
+    font-size: x-large;
+    color: rgb(0, 61, 0);
+    box-shadow: 6px 6px 7px rgb(168, 168, 168);
+    cursor: pointer;
+    outline: none;
+  }
 }
