--- client/src/components/Cards/IrrigationCard/IIrrigationCard.ts
@@ -1,5 +1,0 @@
-interface IIrrigationCardProps {
-  controllerId: string;
-}
-
-export default IIrrigationCardProps;
