+++ client/src/components/Dashboard/DashboardIllustration/IDashboardIllustration.ts
@@ -1,0 +1,5 @@
+interface IDashboardIllustrationProps {
+  controllerId?: string;
+}
+
+export default IDashboardIllustrationProps;
