+++ client/src/components/Dashboard/DashboardInfo/DashboardInfo.module.scss
@@ -6,9 +6,7 @@
   grid-template-rows: 1fr 2fr;
   column-gap: 1rem;
   row-gap: 1rem;
-  grid-template-areas:
-    'smallFirst smallSecond'
-    'big big';
+  grid-area: info;
 }
 
 .card {
