--- client/src/components/DetailBar/DetailBar.tsx
@@ -1,5 +1,0 @@
-const DetailBar: React.FC = () => {
-  return <></>;
-};
-
-export default DetailBar;
