+++ client/src/components/NavBar/NavBar.module.scss
@@ -1,4 +1,4 @@
 .container {
-  grid-area: left-panel;
+  grid-area: logo;
   width: 100%;
 }
