+++ client/src/styles/_vars.scss
@@ -10,3 +10,8 @@ $dark: rgb(98, 98, 91);
 $bg: rgb(252, 240, 200);
 
 $borderRadius: 10px;
+
+$bg: rgb(205, 229, 227);
+$secondaryColor: rgb(61, 92, 98);
+$highlights: rgb(134, 161, 172);
+$highlights2: rgb(147, 201, 187);
